Extracted strings.
