+++ LenovoLegionToolkit.Lib.Automation/Pipeline/Triggers/DisplayOnAutomationPipelineTrigger.cs
@@ -1,0 +1,24 @@
+﻿using System;
+using System.Threading.Tasks;
+using LenovoLegionToolkit.Lib.Automation.Resources;
+using Newtonsoft.Json;
+
+namespace LenovoLegionToolkit.Lib.Automation.Pipeline.Triggers;
+
+public class DisplayOnAutomationPipelineTrigger : IAutomationPipelineTrigger, INativeWindowsMessagePipelineTrigger, IDisallowDuplicatesAutomationPipelineTrigger
+{
+    [JsonIgnore]
+    public string DisplayName => Resource.DisplayOnAutomationPipelineTrigger_DisplayName;
+
+    public Task<bool> IsSatisfiedAsync(IAutomationEvent automationEvent)
+    {
+        var result = automationEvent is NativeWindowsMessageEvent { Message: NativeWindowsMessage.MonitorOn };
+        return Task.FromResult(result);
+    }
+
+    public IAutomationPipelineTrigger DeepCopy() => new DisplayOnAutomationPipelineTrigger();
+
+    public override bool Equals(object? obj) => obj is DisplayOnAutomationPipelineTrigger;
+
+    public override int GetHashCode() => HashCode.Combine(DisplayName);
+}
