+++ LenovoLegionToolkit.Lib.Automation/Pipeline/Triggers/ExternalDisplayConnectedAutomationPipelineTrigger.cs
@@ -7,7 +7,8 @@ namespace LenovoLegionToolkit.Lib.Automation.Pipeline.Triggers;
 
 public class ExternalDisplayConnectedAutomationPipelineTrigger : IAutomationPipelineTrigger, INativeWindowsMessagePipelineTrigger, IDisallowDuplicatesAutomationPipelineTrigger
 {
-    [JsonIgnore] public string DisplayName => Resource.ExternalDisplayConnectedAutomationPipelineTrigger_DisplayName;
+    [JsonIgnore]
+    public string DisplayName => Resource.ExternalDisplayConnectedAutomationPipelineTrigger_DisplayName;
 
     public Task<bool> IsSatisfiedAsync(IAutomationEvent automationEvent)
     {
@@ -18,40 +19,6 @@ public class ExternalDisplayConnectedAutomationPipelineTrigger : IAutomationPipe
     public IAutomationPipelineTrigger DeepCopy() => new ExternalDisplayConnectedAutomationPipelineTrigger();
 
     public override bool Equals(object? obj) => obj is ExternalDisplayConnectedAutomationPipelineTrigger;
-
-    public override int GetHashCode() => HashCode.Combine(DisplayName);
-}
-
-public class DisplayOnAutomationPipelineTrigger : IAutomationPipelineTrigger, INativeWindowsMessagePipelineTrigger, IDisallowDuplicatesAutomationPipelineTrigger
-{
-    [JsonIgnore] public string DisplayName => "When displays turn on";
-
-    public Task<bool> IsSatisfiedAsync(IAutomationEvent automationEvent)
-    {
-        var result = automationEvent is NativeWindowsMessageEvent { Message: NativeWindowsMessage.MonitorOn };
-        return Task.FromResult(result);
-    }
-
-    public IAutomationPipelineTrigger DeepCopy() => new DisplayOnAutomationPipelineTrigger();
-
-    public override bool Equals(object? obj) => obj is DisplayOnAutomationPipelineTrigger;
-
-    public override int GetHashCode() => HashCode.Combine(DisplayName);
-}
-
-public class DisplayOffAutomationPipelineTrigger : IAutomationPipelineTrigger, INativeWindowsMessagePipelineTrigger, IDisallowDuplicatesAutomationPipelineTrigger
-{
-    [JsonIgnore] public string DisplayName => "When displays turn off";
-
-    public Task<bool> IsSatisfiedAsync(IAutomationEvent automationEvent)
-    {
-        var result = automationEvent is NativeWindowsMessageEvent { Message: NativeWindowsMessage.MonitorOff };
-        return Task.FromResult(result);
-    }
-
-    public IAutomationPipelineTrigger DeepCopy() => new DisplayOffAutomationPipelineTrigger();
-
-    public override bool Equals(object? obj) => obj is DisplayOnAutomationPipelineTrigger;
 
     public override int GetHashCode() => HashCode.Combine(DisplayName);
 }
